Split the actor details page into 3 separate folders "crew", "cast" and "Details"
